refactor: Phase 3A - main_test.go大型関数分離で51%削減
main_test.goファイルから最大の関数群を分離し、
ファイルサイズを51%削減して保守性を大幅改善

## 変更内容

### 新規ファイル
- **cmd/beaver/command_integration_test.go**: コマンド統合テスト専用ファイル (1,482行)
  - TestSummarizeCommands (631行): 最大のテスト関数
  - TestRunClassifyIssue/Issues/All (計345行): 分類コマンドテスト群
  - TestRunGenerateWiki/PublishWiki/ListWiki (計156行): Wikiコマンドテスト群
  - TestRunBuildCommand/InitCommand/StatusCommand (計262行): コアコマンドテスト群

- **cmd/beaver/workflow_test.go**: ワークフローテスト専用ファイル (379行)
  - TestInitCommand_FullWorkflow (68行): 初期化ワークフロー
  - TestBuildCommand_FullWorkflow (123行): ビルドワークフロー
  - TestStatusCommand_FullWorkflow (86行): ステータスワークフロー
  - TestFullWorkflowIntegration (81行): 統合ワークフロー

### 変更ファイル
- **cmd/beaver/main_test.go**: 大型テスト関数群を削除
  - ファイルサイズ: ~3,550行 → **1,725行** (51%削減)
  - TestSummarizeCommands等の大型関数を抽出

## 技術仕様
- ✅ 全テストケースを完全保持（機能削除なし）
- ✅ 関数間の依存関係を保持しながら安全に分離
- ✅ インポート最適化とコードフォーマット完了
- ✅ 論理的なファイル分割で関連機能をグループ化

## テスト結果
- ✅ 全テストケースPASS
- ✅ Lint・フォーマット・品質チェック完了
- ✅ 新ファイルの独立性確認済み

## 累積効果 (Phase 1-3A)
- **元のファイル**: ~4,460行
- **現在のファイル**: **1,725行**
- **累積削減**: **2,735行** (61%削減)
- **新テストファイル**: 5ファイル追加で機能別整理完了
